+++ src/bad_credentials/index.d.ts
@@ -1,8 +1,8 @@
-import type { error_o_T } from '../_types'
+import type { argument__error_o_T, error_o_error_T, error_o_T } from '../_types'
 import type { HttpError } from '../http_error'
 export declare function bad_credentials_error_(
 	message?:string,
-	bad_credentials_error_o?:bad_credentials_error_o_T
+	bad_credentials_error_o?:argument__bad_credentials_error_o_T
 ):BadCredentialsError
 /**
  * Throws a Bad Credentials error (HTTP 401)
@@ -11,9 +11,15 @@ export declare function bad_credentials_error_(
  */
 export declare function bad_credentials__throw(
 	message?:string,
-	bad_credentials_error_o?:error_o_T
+	bad_credentials_error_o?:argument__bad_credentials_error_o_T
 ):void
-export declare class BadCredentialsError extends HttpError implements bad_credentials_error_o_T {}
+export declare class BadCredentialsError
+	extends HttpError
+	implements bad_credentials_error_o_error_T {}
 export interface bad_credentials_error_o_T extends error_o_T {
 	jwt_token?:string
 }
+export interface argument__bad_credentials_error_o_T
+	extends bad_credentials_error_o_T, argument__error_o_T {}
+export interface bad_credentials_error_o_error_T
+	extends error_o_error_T, bad_credentials_error_o_T {}