+++ src/bad_credentials/index.js
@@ -1,9 +1,9 @@
 import { assign } from '@ctx-core/object'
 import { HttpError } from '../http_error/index.js'
-/** @typedef {import('./index.d.ts').bad_credentials_error_o_T}bad_credentials_error_o_T */
+/** @typedef {import('./index.d.ts').argument__bad_credentials_error_o_T}argument__bad_credentials_error_o_T */
 /**
  * @param {string}[message]
- * @param {bad_credentials_error_o_T}[bad_credentials_error_o]
+ * @param {argument__bad_credentials_error_o_T}[bad_credentials_error_o]
  * @private
  */
 export function bad_credentials_error_(
@@ -15,7 +15,7 @@ export function bad_credentials_error_(
 /**
  * Throws a Bad Credentials error (HTTP 401)
  * @param {string}[message]
- * @param {bad_credentials_error_o_T}[bad_credentials_error_o]
+ * @param {argument__bad_credentials_error_o_T}[bad_credentials_error_o]
  * @example
  * bad_credentials__throw(ctx) // Unauthorized
  */