+++ src/bad_gateway/index.d.ts
@@ -1,8 +1,8 @@
-import type { error_o_T } from '../_types'
+import type { argument__error_o_T, error_o_error_T } from '../_types'
 import type { HttpError } from '../http_error'
 export declare function bad_gateway_error_(
 	message?:string,
-	error_o?:error_o_T
+	error_o?:argument__error_o_T
 ):BadGatewayError
 /**
  * Throws a bad_gateway error (HTTP 502)
@@ -11,6 +11,8 @@ export declare function bad_gateway_error_(
  */
 export declare function bad_gateway__throw(
 	message?:string,
-	error_o?:error_o_T
+	error_o?:argument__error_o_T
 ):void
-export declare class BadGatewayError extends HttpError {}
+export declare class BadGatewayError
+	extends HttpError
+	implements error_o_error_T {}