+++ src/console__error/index.d.ts
@@ -1,7 +1,7 @@
-import type { error_o_T } from '../_types'
+import type { error_o_error_T } from '../_types'
 /**
  * Logs the given error_o to stderr
  */
 export declare function console__error(
-	error_o:error_o_T
+	error_o_error:error_o_error_T
 ):void