+++ src/error__print/index.d.ts
@@ -1,5 +1,5 @@
-import type { error_o_T } from '../_types'
+import type { error_o_error_T } from '../_types'
 /**
  * Prints the given `error_o` to stderr
  */
-export declare function error__print(error_o:error_o_T):void
+export declare function error__print(error_o_error:error_o_error_T):void