+++ src/error_o/index.js
@@ -1,11 +1,11 @@
 import { assign } from '@ctx-core/object'
-import { console__error } from '../console__error/index.js'
-/** @typedef {import('../_types').error_o_T}error_o_T */
+/** @typedef {import('../_types').argument__error_o_T}argument__error_o_T */
+/** @typedef {import('../_types').error_o_error_T}error_o_error_T */
 /**
- * error_o_T factory function
+ * argument__error_o_T factory function
  * @param {string}[message]
- * @param {error_o_T}[error_o]
- * @returns {Error&error_o_T}
+ * @param {argument__error_o_T}[error_o]
+ * @returns {error_o_error_T}
  */
 export function error_o_(message, error_o) {
 	const error = new Error(message)
@@ -14,7 +14,7 @@ export function error_o_(message, error_o) {
 /**
  * Throws an error_o
  * @param {string}[message]
- * @param {error_o_T}[error_o]
+ * @param {argument__error_o_T}[error_o]
  * @throws Decorate & throw error given by the arguments.
  */
 export function error_o__throw(message, error_o) {