+++ src/http_error/index.d.ts
@@ -1,7 +1,7 @@
-import type { error_o_T } from '../_types'
+import type { argument__error_o_T, error_o_error_T } from '../_types'
 export declare function http_error_(
 	message?:string,
-	error_o?:error_o_T
+	error_o?:argument__error_o_T
 ):HttpError
 /**
  * HttpError
@@ -10,6 +10,8 @@ export declare function http_error_(
  */
 export declare function http_error__throw(
 	message?:string,
-	error_o?:error_o_T
+	error_o?:argument__error_o_T
 ):void
-export declare class HttpError extends Error implements error_o_T {}
+export declare class HttpError
+	extends Error
+	implements error_o_error_T {}