+++ src/invalid_state/index.d.ts
@@ -1,8 +1,8 @@
-import type { error_o_T } from '../_types'
+import type { argument__error_o_T, error_o_error_T, error_o_T } from '../_types'
 import type { HttpError } from '../http_error'
 export declare function invalid_state_error_(
 	message?:string,
-	invalid_state_error_o?:invalid_state_error_o_T
+	invalid_state_error_o?:argument__invalid_state_error_o_T
 ):InvalidStateError
 /**
  * Throws an invalid_state error (HTTP 500)
@@ -11,9 +11,11 @@ export declare function invalid_state_error_(
  */
 export declare function invalid_state__throw(
 	message?:string,
-	invalid_state_error_o?:invalid_state_error_o_T
+	invalid_state_error_o?:argument__invalid_state_error_o_T
 ):void
-export declare class InvalidStateError extends HttpError implements invalid_state_error_o_T {
+export declare class InvalidStateError
+	extends HttpError
+	implements invalid_state_error_o_error_T {
 	key?:string
 	reason?:string
 }
@@ -21,3 +23,8 @@ export interface invalid_state_error_o_T extends error_o_T {
 	key?:string
 	reason?:string
 }
+export interface argument__invalid_state_error_o_T
+	extends invalid_state_error_o_T, argument__error_o_T {
+}
+export interface invalid_state_error_o_error_T
+	extends invalid_state_error_o_T, error_o_error_T {}