+++ src/missing_argument/index.d.ts
@@ -1,4 +1,4 @@
-import type { error_o_T } from '../_types'
+import type { argument__error_o_T, error_o_error_T, error_o_T } from '../_types'
 import type { HttpError } from '../http_error'
 export declare function missing_argument_error_(
 	message?:string,
@@ -13,7 +13,13 @@ export declare function missing_argument__throw(
 	message?:string,
 	missing_argument_error_o?:missing_argument_error_o_T
 ):void
-export declare class MissingArgumentError extends HttpError implements missing_argument_error_o_T {}
+export declare class MissingArgumentError
+	extends HttpError
+	implements missing_argument_error_o_error_T {}
 export interface missing_argument_error_o_T extends error_o_T {
 	key?:string
 }
+export interface argument__missing_argument_error_o_T
+	extends missing_argument_error_o_T, argument__error_o_T {}
+export interface missing_argument_error_o_error_T
+	extends error_o_error_T, missing_argument_error_o_T {}