+++ src/missing_argument/index.js
@@ -1,22 +1,22 @@
 import { assign } from '@ctx-core/object'
 import { HttpError } from '../http_error/index.js'
-/** @typedef {import('./index.d.ts').missing_argument_error_o_T}missing_argument_error_o_T */
+/** @typedef {import('./index.d.ts').argument__missing_argument_error_o_T}argument__missing_argument_error_o_T */
 /**
  * @param {string}[message]
- * @param {missing_argument_error_o_T}[missing_argument_error_o]
+ * @param {argument__missing_argument_error_o_T}[missing_argument_error_o]
  * @returns {MissingArgumentError}
  * @private
  */
 export function missing_argument_error_(
 	message,
- 	missing_argument_error_o
+	missing_argument_error_o
 ) {
 	return assign(new MissingArgumentError(message), missing_argument_error_o)
 }
 /**
  * Throws a missing_argument error (HTTP 500)
  * @param {string}[message]
- * @param {missing_argument_error_o_T}[missing_argument_error_o]
+ * @param {argument__missing_argument_error_o_T}[missing_argument_error_o]
  * @example
  * missing_argument__throw('foobar: not defined', { key: 'foobar' })
  */