+++ src/unauthorized/index.d.ts
@@ -1,4 +1,4 @@
-import type { error_o_T } from '../_types'
+import type { error_o_error_T, error_o_T } from '../_types'
 import type { HttpError } from '../http_error'
 export declare function unauthorized_error_(
 	message?:string,
@@ -13,4 +13,6 @@ export declare function unauthorized__throw(
 	message?:string,
 	error_o?:error_o_T
 ):void
-export declare class UnauthorizedError extends HttpError {}
+export declare class UnauthorizedError
+	extends HttpError
+	implements error_o_error_T {}